(1.1) Fixed compile warning - unsigned/signed mismatch.
git-svn-id: http://svn.gplates.org/projects/gplates/branches/1.1@11691 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/opengl/GLMultiResolutionFilledPolygons.cc
@@ -120,7 +120,8 @@ GPlatesOpenGL::GLMultiResolutionFilledPolygons::GLMultiResolutionFilledPolygons(
 	// But it can't be larger than the maximum texture dimension for the current system.
 	//
 	// NOTE: An OpenGL context must be active at this point.
-	if (d_polygon_stencil_texel_dimension > GLContext::get_texture_parameters().gl_max_texture_size)
+	if (d_polygon_stencil_texel_dimension >
+		boost::numeric_cast<unsigned int>(GLContext::get_texture_parameters().gl_max_texture_size))
 	{
 		d_polygon_stencil_texel_dimension = GLContext::get_texture_parameters().gl_max_texture_size;
 	}
